master + join password limit

+++ app/src/main/java/com/example/umc_hackathon/post/FormDetailActivity.kt
@@ -1,20 +1,31 @@
 package com.example.umc_hackathon
 
+import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.view.View
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
+import androidx.core.content.ContextCompat.startActivity
 import com.example.umc_hackathon.auth.AuthActivity
+import com.example.umc_hackathon.auth.AuthService
+import com.example.umc_hackathon.auth.ReAccessTokenResponse
+import com.example.umc_hackathon.auth.ReAccessTokenView
 import com.example.umc_hackathon.databinding.ActivityFormDetailBinding
 import com.example.umc_hackathon.post.*
 import com.example.umc_hackathon.post.result.ResultActivity
 import com.example.umc_hackathon.survey.FormInputActivity
+import com.example.umc_hackathon.survey.ModifyActivity
 
-class FormDetailActivity : AppCompatActivity(), PostDetailView {
+class FormDetailActivity : AppCompatActivity(), PostDetailView, ReAccessTokenView {
 
     private var postId: Long = 0L
     private lateinit var postTitle: String
+    private var postDeadline: Int = 0
+    private var postUserId: Long = 0
+    private lateinit var postContent: String
     private lateinit var binding: ActivityFormDetailBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -48,13 +59,33 @@ class FormDetailActivity : AppCompatActivity(), PostDetailView {
             likePost()
         }
         binding.formDetailDeleteBtn.setOnClickListener {
-            deletePost()
+            AlertDialog.Builder(this).run {
+                setTitle("설문을 삭제하시겠습니까?")
+                setPositiveButton("네, 삭제하겠습니다", DialogInterface.OnClickListener { dialogInterface, i ->
+                    deletePost()
+                })
+                setNegativeButton("아니요", null)
+                show()
+            }
         }
+
         binding.formDetailResultBtn.setOnClickListener {
             val intent = Intent(this, ResultActivity::class.java)
             intent.addFlags (Intent.FLAG_ACTIVITY_NO_ANIMATION)
             intent.putExtra("postId", postId)
             intent.putExtra("postTitle", postTitle)
+            startActivity(intent)
+            finish()
+        }
+
+        binding.formDetailModifyBtn.setOnClickListener {
+            val intent = Intent(this, ModifyActivity::class.java)
+            intent.addFlags (Intent.FLAG_ACTIVITY_NO_ANIMATION)
+            intent.putExtra("postId", postId)
+            intent.putExtra("postTitle", postTitle)
+            intent.putExtra("postContent", postContent)
+            intent.putExtra("postDeadline", postDeadline)
+            intent.putExtra("postUserId", postUserId)
             startActivity(intent)
             finish()
         }
@@ -65,13 +96,13 @@ class FormDetailActivity : AppCompatActivity(), PostDetailView {
 
         val postService = PostService()
         postService.setPostDetailView(this)
-        postService.getPostDetail(postId, getJwt().toString())
+        postService.getPostDetail(postId, getAccessToken().toString(), getRefreshToken().toString())
     }
 
     private fun likePost() {
         val postService = PostService()
         postService.setPostDetailView(this)
-        postService.likePost(postId, getJwt().toString())
+        postService.likePost(postId, getAccessToken().toString(), getRefreshToken().toString())
     }
 
     private fun dislikePost() {
@@ -86,18 +117,50 @@ class FormDetailActivity : AppCompatActivity(), PostDetailView {
         postService.deletePost(postId, getJwt().toString())
     }
 
+    private fun saveAccessToken(accessToken: String) {
+        val spf = getSharedPreferences("auth", AppCompatActivity.MODE_PRIVATE)
+        val editor = spf.edit()
+
+        editor.putString("accessToken", accessToken)
+        editor.apply()
+
+        Log.d("엑세스토근", "세이브")
+    }
+
+    private fun getReAccessToken() {
+        val authService = AuthService()
+        authService.setReAccessTokenView(this)
+        authService.getReAccessToken(getAccessToken().toString(), getRefreshToken().toString())
+    }
+
     private fun getJwt(): String? {
         val spf = this.getSharedPreferences("auth", AppCompatActivity.MODE_PRIVATE)
         return spf!!.getString("jwt", "")
     }
 
+    private fun getAccessToken(): String? {
+        val spf = this.getSharedPreferences("auth", AppCompatActivity.MODE_PRIVATE)
+        return spf!!.getString("accessToken", "")
+    }
+
+    private fun getRefreshToken(): String? {
+        val spf = this.getSharedPreferences("auth", AppCompatActivity.MODE_PRIVATE)
+        return spf!!.getString("refreshToken", "")
+    }
+
     override fun onGetPostDetailSuccess(result: PostDetailResult) {
         postTitle = result.title
+        postDeadline = result.dday
+        postContent = result.content
+        postUserId = result.postUserId
+
         binding.formDetailTitleTv.text = postTitle
         binding.formDetailInfoTv.text = result.content
         binding.formDetailItemCountTv.text = result.qCount.toString() + "개의 항목"
 
         if (result.status == "ACTIVE") {
+            binding.formDetailModifyBtn.visibility = View.VISIBLE
+
             if (result.dday == 0) {
                 binding.formDetailItemDeadlineTv.text = "D - DAY"
             }
@@ -137,11 +200,16 @@ class FormDetailActivity : AppCompatActivity(), PostDetailView {
         Log.d("PostDetail / ", "상세페이지를 불러오는데 성공했습니다")
     }
 
-    override fun onGetPostDetailFailure() {
-        Log.d("PostDetail / ", "상세페이지를 불러오는데 실패했습니다")
-        val intent = Intent(this, AuthActivity::class.java)
-        startActivity(intent)
-        finish()
+    override fun onGetPostDetailFailure(result: PostDetailResponse) {
+        Log.d("PostDetail / ", "상세페이지를 불러오는데 실패했습니다" + result.code)
+
+        if(result.code == 2002) {
+            getReAccessToken()
+        } else {
+            val intent = Intent(this, AuthActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
     }
 
     override fun onLikeSuccess() {
@@ -163,11 +231,21 @@ class FormDetailActivity : AppCompatActivity(), PostDetailView {
     }
 
     override fun onDeleteSuccess() {
-        //삭제 알림창 띄우기
+        Toast.makeText(this, "설문 삭제를 성공했습니다.", Toast.LENGTH_SHORT).show()
     }
 
     override fun onDeleteFailure(result: StringResultResponse) {
+        Toast.makeText(this, "설문 삭제를 실패했습니다.", Toast.LENGTH_SHORT).show()
         Log.d("deletePost()", " 실패 / " + result.message)
+    }
+
+    override fun onGetReAccessTokenSuccess(res: ReAccessTokenResponse) {
+        Log.d("액세스토근", "을 재발급했습니다.")
+        saveAccessToken(res.result)
+    }
+
+    override fun onGetReAccessTokenFailure() {
+        Log.d("onGetReAccessTokenFailure()", " 실패 / ")
     }
 
 }
